+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- AI Guard: `messages` parameter is no longer a generic. A new `Message` model
+  has been introduced, and `messages` is now a `Sequence[Message]`.
+
 ## 6.1.1 - 2025-05-12
 
 ### Fixed